History no longer exists, all will be saved in the same directory of outputs, by determining in which simulation the networks are recorded
since we run multiple time each simulation to get an average of the results, now this number is mentioned on the saved files

+++ Functions/setting_initial_values.cpp
@@ -1,6 +1,6 @@
 #include "../Headers/Genome.hpp"
 
-void genome::Setting_initial_values(int N, int st, int Nn_net, float Mute_R, double env_ref, int func_env, int RUN)
+void genome::Setting_initial_values(int N, int st, int Nn_net, float Mute_R, float env_ref, int func_env, int RUN)
 {
     int number_networks = Nn_net;
